src/requirements.txt
Django==4.2
djangorestframework==3.14
neomodel==5.1.0
django-neomodel==0.1.1
datasets==2.16.1
gunicorn==21.2.0
drf-spectacular==0.27.0
